+++ src-docs/src/views/markdown_editor/mardown_editor_example.js
@@ -16,10 +16,52 @@ import { Link } from 'react-router-dom';
 import MarkdownEditor from './markdown_editor';
 const markdownEditorSource = require('!!raw-loader!./markdown_editor');
 const markdownEditorHtml = renderToHtml(MarkdownEditor);
+const markdownEditorSnippet = `<EuiMarkdownEditor
+  value={value}
+  onChange={setValue}
+/>`;
 
 import MarkdownEditorErrors from './markdown_editor_errors';
 const markdownEditorErrorsSource = require('!!raw-loader!./markdown_editor_errors');
 const markdownEditorErrorsHtml = renderToHtml(MarkdownEditorErrors);
+const markdownEditorErrorsSnippet = `<EuiMarkdownEditor
+  value={value}
+  onChange={setValue}
+  onParse={onParse}
+  errors={messages}
+/>`;
+
+import MarkdownEditorHeight from './markdown_editor_height';
+const markdownEditorHeightSource = require('!!raw-loader!./markdown_editor_height');
+const markdownEditorHeightHtml = renderToHtml(MarkdownEditorHeight);
+const markdownEditorHeightSnippet = [
+  `// Custom height with auto-expanding preview
+<EuiMarkdownEditor
+  value={value}
+  onChange={setValue}
+  height={200}
+/>`,
+  `// Height set to full
+<EuiMarkdownEditor
+  value={value}
+  onChange={setValue}
+  height="full"
+/>`,
+  `// Custom height with no auto-expanding preview
+<EuiMarkdownEditor
+  value={value}
+  onChange={setValue}
+  height={200}
+  autoExpandPreview={false}
+/>`,
+  `// Custom height and custom max height
+<EuiMarkdownEditor
+  value={value}
+  onChange={setValue}
+  height={200}
+  maxHeight={300}
+/>`,
+];
 
 export const MarkdownEditorExample = {
   title: 'Markdown editor',
@@ -66,6 +108,7 @@ export const MarkdownEditorExample = {
       props: {
         EuiMarkdownEditor,
       },
+      snippet: markdownEditorSnippet,
       demo: <MarkdownEditor />,
     },
     {
@@ -93,7 +136,47 @@ export const MarkdownEditorExample = {
       props: {
         EuiMarkdownEditor,
       },
+      snippet: markdownEditorErrorsSnippet,
       demo: <MarkdownEditorErrors />,
+    },
+    {
+      source: [
+        {
+          type: GuideSectionTypes.JS,
+          code: markdownEditorHeightSource,
+        },
+        {
+          type: GuideSectionTypes.HTML,
+          code: markdownEditorHeightHtml,
+        },
+      ],
+      title: 'Controlling the height',
+      text: (
+        <>
+          <p>
+            The <EuiCode>height</EuiCode> prop allows you to control the height
+            of the <strong>EuiMarkdownEditor</strong>. You can set the{' '}
+            <EuiCode>height</EuiCode> in pixels or pass{' '}
+            <EuiCode>&quot;full&quot;</EuiCode> to allow the{' '}
+            <strong>EuiMarkdownEditor</strong> to fill the height of its
+            container. By default, the <EuiCode>autoExpandPreview</EuiCode> prop
+            is set to <EuiCode>true</EuiCode>. This means that the preview{' '}
+            <EuiCode>height</EuiCode> is automatically adjusted to fit all the
+            content and avoid a scrollbar.
+          </p>
+          <p>
+            You can also control the <EuiCode>maxHeight</EuiCode> of the{' '}
+            editor/preview area. This prop only works when the{' '}
+            <EuiCode>height</EuiCode> is not set to{' '}
+            <EuiCode>&quot;full&quot;</EuiCode>.
+          </p>
+        </>
+      ),
+      props: {
+        EuiMarkdownEditor,
+      },
+      snippet: markdownEditorHeightSnippet,
+      demo: <MarkdownEditorHeight />,
     },
   ],
 };